+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * perf: warmable cache
 * fix(`CacheWarmer`): use `symfony/filesystem` to dump file
+* fix(`PHPStan`): fix mapping metadata generation determinism
 
 ## 1.9.4