+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * fix(`CacheWarmer`): use `symfony/filesystem` to dump file
 * fix(`PHPStan`): fix mapping metadata generation determinism
 * deps: remove phpstan unused public
+* feat: attributes inheritance is now preserved
 
 ## 1.9.4